refactor: 추천 관련 화면 패키지 이동 및 UI 컴포넌트 경로 구조 재구성
